cashbook_system/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.36
Pillow==10.4.0
openpyxl==3.1.2
reportlab==4.0.7
xlsxwriter==3.2.0